+++ .jetclient/jetclient.md
@@ -1,6 +1,8 @@
 ```toml
 name = 'Project 2-2'
 id = '3db97993-1289-45cb-a6e2-3f369c60d81e'
+name = 'Inventory'
+id = '2b5db64d-efb1-46ca-bda3-93d6090bd7d1'
 
 [[environmentGroups]]
 name = 'Default'
