--- client/src/pages/anuththara/ProductList.js
@@ -1,59 +1,0 @@
-import React, { useEffect, useState } from 'react';
-import axios from 'axios';
-
-const ProductList = () => {
-    const [products, setProducts] = useState([]);
-    const [searchTerm, setSearchTerm] = useState("");
-
-    useEffect(() => {
-        const fetchProducts = async () => {
-            const response = await axios.get('/api/products');
-            setProducts(response.data);
-        };
-        fetchProducts();
-    }, []);
-
-    const filteredProducts = products.filter(product =>
-        product.title.toLowerCase().includes(searchTerm.toLowerCase())
-    );
-
-    return (
-        <div className="container mx-auto">
-            <input
-                type="text"
-                placeholder="Search products..."
-                value={searchTerm}
-                onChange={(e) => setSearchTerm(e.target.value)}
-                className="border p-2 m-4"
-            />
-            <table className="table-auto w-full">
-                <thead>
-                <tr>
-                    <th>Title</th>
-                    <th>Price</th>
-                    <th>Category</th>
-                    <th>Stock</th>
-                    <th>Actions</th>
-                </tr>
-                </thead>
-                <tbody>
-                {filteredProducts.map(product => (
-                    <tr key={product.id}>
-                        <td>{product.title}</td>
-                        <td>{product.price}</td>
-                        <td>{product.category}</td>
-                        <td>{product.quantity}</td>
-                        <td>
-                            <button onClick={() => handleEdit(product.id)}>Edit</button>
-                            <button onClick={() => handleDelete(product.id)}>Delete</button>
-                            <button onClick={() => handleView(product.id)}>View</button>
-                        </td>
-                    </tr>
-                ))}
-                </tbody>
-            </table>
-        </div>
-    );
-};
-
-export default ProductList;
