+++ client/src/pages/anuththara/StatsPage.js
@@ -1,6 +1,8 @@
-import React, { useState, useEffect } from 'react';
+import React, { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import axios from 'axios';
+import Sidebar from '../com/admindash'; // Import your Sidebar component
+import homepic7 from "../../images/f.jpg";
 
 const StatsPage = () => {
     const [stats, setStats] = useState({
@@ -13,7 +15,6 @@ const StatsPage = () => {
     useEffect(() => {
         const fetchStats = async () => {
             try {
-                // Fetch stats from the backend server running on port 3001
                 const response = await axios.get('http://localhost:3001/api/products/stats');
                 setStats(response.data);
             } catch (error) {
@@ -22,7 +23,7 @@ const StatsPage = () => {
         };
 
         fetchStats();
-    }, []);
+    }, []); // Empty dependency array ensures this runs only once
 
     return (
         <div className="p-8 space-y-6">
@@ -42,7 +43,7 @@ const StatsPage = () => {
                 </div>
             </div>
             <button
-                onClick={() => navigate('/products')}
+                onClick={() => navigate('/ProductList')}
                 className="mt-6 px-4 py-2 bg-blue-500 text-white rounded-lg shadow hover:bg-blue-600"
             >
                 Go to Product List
@@ -51,4 +52,57 @@ const StatsPage = () => {
     );
 };
 
-export default StatsPage;
+const Layout = () => {
+    const navigate = useNavigate();
+
+    useEffect(() => {
+        const checkAdmin = async () => {
+            try {
+                const response = await fetch('http://localhost:3001/api/user/admin', {
+                    credentials: 'include', // Include credentials with the request
+                });
+
+                if (response.status === 403 || response.status === 401) {
+                    navigate('/'); // Redirect if not authorized
+                    return;
+                }
+
+                const data = await response.json();
+                if (!data.isAdmin) {
+                    navigate('/'); // Redirect if the user is not an admin
+                }
+            } catch (error) {
+                console.error('Error checking user role:', error);
+                navigate('/'); // Redirect in case of an error
+            }
+        };
+
+        checkAdmin();
+    }, [navigate]);
+
+    return (
+        <div className="flex h-screen">
+            <div className="w-[20%] h-full text-white"
+                 style={{
+                     backgroundImage: `url(${homepic7})`,
+                     backgroundSize: 'cover',
+                     backgroundPosition: 'center',
+                     backgroundRepeat: 'no-repeat',
+                 }}>
+                <Sidebar />
+            </div>
+            <div className="w-[80%] h-full bg-pink-500 p-4 julius-sans-one-regular"
+                 style={{
+                     backgroundImage: `url(${homepic7})`,
+                     backgroundSize: 'cover',
+                     backgroundPosition: 'center',
+                     backgroundRepeat: 'no-repeat',
+                 }}>
+                {/* You can place other content or components here */}
+                <StatsPage />
+            </div>
+        </div>
+    );
+};
+
+export default Layout;
