+++ server/package.json
@@ -15,6 +15,7 @@
     "@google/generative-ai": "^0.17.1",
     "axios": "^1.7.5",
     "bcrypt": "^5.1.1",
+    "blob-stream": "^0.1.3",
     "body-parser": "^1.20.2",
     "cors": "^2.8.5",
     "dotenv": "^16.3.1",
@@ -22,11 +23,14 @@
     "express": "^4.18.2",
     "express-async-errors": "^3.1.1",
     "express-session": "^1.17.3",
+    "file-saver": "^2.0.5",
     "js-cookie": "^3.0.5",
+    "multer": "^1.4.5-lts.1",
     "passport": "^0.7.0",
     "passport-google-oauth2": "^0.2.0",
     "passport-google-oauth20": "^2.0.0",
     "passport-local": "^1.0.0",
+    "pdfkit": "^0.15.0",
     "pg": "^8.11.3"
   }
 }
